platform/cli/requirements.txt
# Noesis CLI Requirements
click>=8.0.0
PyYAML>=6.0
